+++ acceptance-tests/src/test/java/tech/pegasys/ethsigner/tests/SignTransactionAcceptanceTest.java
@@ -13,6 +13,7 @@
 package tech.pegasys.ethsigner.tests;
 
 import static org.assertj.core.api.Assertions.assertThat;
+import static tech.pegasys.ethsigner.tests.WaitUtils.waitFor;
 
 import tech.pegasys.ethsigner.tests.dsl.Accounts;
